fixed weird problem, wtf
before this change, the bot was always heading to (0,0), instead
of heading towards the next available food. This seems to be related
to world.cells being a defaultdict(Cell) (and not an ordinary dict).
Also, changing the dict index `eater_id` to something nonexistent like
`1`, the bug vanished as well.